+++ my3dgsapp/android/app/src/main/java/tw/threewa/my3dgsapp/MainActivity.java
@@ -1,0 +1,5 @@
+package tw.threewa.my3dgsapp;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}